+++ .gitignore
@@ -8,3 +8,4 @@ sample_feed.py
 test.py
 *.ipynb
 .ipynb_checkpoints/
+*.db
