+++ main.py
@@ -1,6 +1,4 @@
-from market_data import app
-from market_data import scheduler, marketFeedTicker, marketFeedQuote
+from api import app
 
 if __name__ == "__main__":
-    scheduler.start()
     app.run(debug=True)
