--- market_data/market_feed/__init__.py
@@ -1,2 +1,0 @@
-from .dhan_market_feed import *
-from .template import *
