--- market_data/market_feed/template.py
@@ -1,10 +1,0 @@
-class Template:
-    def __init__(self, analyser):
-        self.analyser = analyser
-        pass
-
-    def connect():
-        pass
-
-    def parse():
-        pass
