+++ market_data/models.py
@@ -1,4 +1,3 @@
-from market_data import db
 from dataclasses import dataclass
 import datetime
 
@@ -48,23 +47,3 @@ class MarketDepthData:
     ask_price: list
     bid_orders: list
     ask_orders: list
-
-
-class APIKey(db.Model):
-    id = db.Column(db.Integer, primary_key=True)
-    key = db.Column(db.String(500), nullable=False)
-    secret = db.Column(db.String(100), nullable=False)
-    expiry = db.Column(db.DateTime, nullable=True)
-    platform = db.Column(db.String(100), nullable=False)
-
-    def __repr__(self):
-        return f"APIKey(key={self.key}, secret={self.secret})"
-
-
-class Symbol(db.Model):
-    id = db.Column(db.Integer, primary_key=True)
-    symbol = db.Column(db.String(100), nullable=False)
-    exchange = db.Column(db.String(100), nullable=False)
-
-    def __repr__(self):
-        return f"Symbol(symbol={self.symbol}, exchange={self.exchange})"
