+++ requirements.txt
@@ -16,6 +16,7 @@ Jinja2==3.1.3
 MarkupSafe==2.1.5
 numpy==1.26.4
 pandas==2.2.2
+psycopg2-binary==2.9.9
 python-dateutil==2.9.0.post0
 pytz==2024.1
 redis==5.0.3
